+++ lib/bali/rails/active_record.rb
@@ -5,4 +5,6 @@ require 'active_support/lazy_load_hooks'
 ActiveSupport.on_load :active_record do
   require "bali"
   ::ActiveRecord::Base.send :include, Bali::Authorizer
+  ::ActiveRecord::Base.send :extend, Bali::Statics::Authorizer
+  ::ActiveRecord::Base.send :extend, Bali::Statics::ActiveRecord
 end